Imu read Kernel working

+++ creator/lsm9ds1.h
@@ -166,21 +166,21 @@ class LSM9DS1 {
   // DPS. This function relies on gScale and gRes being correct.
   // Input:
   //	- gyro = A signed 16-bit raw reading from the gyroscope.
-  float calcGyro(int16_t gyro);
+  int calcGyro(int16_t gyro);
 
   // calcAccel() -- Convert from RAW signed 16-bit value to gravity (g's).
   // This function reads in a signed 16-bit value and returns the scaled
   // g's. This function relies on aScale and aRes being correct.
   // Input:
   //	- accel = A signed 16-bit raw reading from the accelerometer.
-  float calcAccel(int16_t accel);
+  int calcAccel(int16_t accel);
 
   // calcMag() -- Convert from RAW signed 16-bit value to Gauss (Gs)
   // This function reads in a signed 16-bit value and returns the scaled
   // Gs. This function relies on mScale and mRes being correct.
   // Input:
   //	- mag = A signed 16-bit raw reading from the magnetometer.
-  float calcMag(int16_t mag);
+  int calcMag(int16_t mag);
 
   // setGyroScale() -- Set the full-scale range of the gyroscope.
   // This function can be called to set the scale of the gyroscope to
